Move pkg/util/logs to staging and update ref
